+++ CHANGELOG.rst
@@ -7,6 +7,13 @@ Change Log
 ----------
 
 
+8.7.2
+=====
+
+* Changes to itemize SMaHT submission ingestion create/update/diff situation (portal_object_utils).
+* Changes to structured_data to handle property deletes (portal_object_utils).
+
+
 8.7.1
 =====